requirements.txt
customtkinter
requests
psycopg2-binary
tkcalendar
python-dateutil
pytz